feat(FFmpegBuilder): support -bsf:v and -bsf:a for bitstream filters
In order to add the possibility to use Bitstream filter for audio and video, some attribute have been added to the OutputBuilder so the parameters are integrated in the command line. This modification allow in my project (in link in this PR) to use the 'aac_adtstoasc' (https://www.ffmpeg.org/ffmpeg-bitstream-filters.html#aac_005fadtstoasc) which allow to merge a AAC and MP4 video together without re-encoding. This format t is the only one usable in web environnement without major re-enconding.

  The support of this bitstream filters are quiet simple, and they can be complexified to support parameters in future release.