feat(builder): set log level programmatically with the builder
I'v create a enum with all the possible value and define a default value in the builder to the previous one (ERROR). The user can set verbosity with method `setVerbosity(Verbosity v)`. This will help us to develop the issue #21 because we need a certain level of information to display / send the progresion of a ffmpeg operation